requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.5.0
matplotlib>=3.5.0
tqdm>=4.64.0
python-dotenv>=1.0.0

# NLP dependencies
spacy>=3.5.0
nltk>=3.8.0
sentence-transformers>=2.2.0
transformers>=4.28.0
rouge-score>=0.1.2

# Vision dependencies
pillow>=9.4.0
torch>=2.0.0
torchvision>=0.15.0

# Install spaCy model
# Run after installing dependencies:
# python -m spacy download en_core_web_sm

# To install dependencies using uv:
# uv pip install -r requirements.txt 